added version, model, vendor for app-source and activeappsource
