+++ java-sdk/radar-schemas-tools/src/main/java/org/radarcns/schema/specification/AppSource.java
@@ -2,12 +2,24 @@ package org.radarcns.schema.specification;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
+import java.util.Objects;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 import static org.radarcns.schema.util.Utils.expandClass;
 
 public abstract class AppSource<T extends DataTopic> extends DataProducer<T> {
     @JsonProperty("app_provider")
     private String appProvider;
+
+    @JsonProperty
+    private String vendor;
+
+    @JsonProperty
+    private String model;
+
+    @JsonProperty
+    private String version;
 
     @JsonSetter
     @SuppressWarnings("PMD.UnusedPrivateMethod")
@@ -17,5 +29,38 @@ public abstract class AppSource<T extends DataTopic> extends DataProducer<T> {
 
     public String getAppProvider() {
         return appProvider;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public String getVendor() {
+        return vendor;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        AppSource provider = (AppSource) o;
+        return Objects.equals(appProvider, provider.appProvider)
+                && Objects.equals(version, provider.version)
+                && Objects.equals(model, provider.model)
+                && Objects.equals(vendor, provider.vendor)
+                && Objects.equals(getData(), provider.getData());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(appProvider, vendor, model, version, getData());
     }
 }
