+++ java-sdk/radar-schemas-tools/src/main/java/org/radarcns/schema/specification/DataProducer.java
@@ -26,19 +26,12 @@ public abstract class DataProducer<T extends DataTopic> {
     @JsonProperty
     private List<String> labels;
 
-    @JsonProperty
-    private String version;
-
     public String getName() {
         return name;
     }
 
     public String getDoc() {
         return doc;
-    }
-
-    public String getVersion() {
-        return version;
     }
 
     @NotNull
@@ -75,12 +68,11 @@ public abstract class DataProducer<T extends DataTopic> {
         DataProducer producer = (DataProducer) o;
         return Objects.equals(name, producer.name)
                 && Objects.equals(doc, producer.doc)
-                && Objects.equals(version , producer.version)
                 && Objects.equals(getData(), producer.getData());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name, doc, getData(), version);
+        return Objects.hash(name, doc, getData());
     }
 }
