+++ java-sdk/radar-schemas-tools/src/main/java/org/radarcns/schema/specification/passive/PassiveSource.java
@@ -33,12 +33,6 @@ public class PassiveSource extends AppSource<PassiveDataTopic> {
         EMPATICA_E4, PEBBLE_2, ANDROID_PHONE, BIOVOTION_VSM1
     }
 
-    @JsonProperty @NotBlank
-    private String vendor;
-
-    @JsonProperty @NotBlank
-    private String model;
-
     @JsonProperty @NotEmpty
     private List<PassiveDataTopic> data;
 
@@ -53,15 +47,15 @@ public class PassiveSource extends AppSource<PassiveDataTopic> {
     }
 
     public String getName() {
-        return vendor + '_' + model;
+        return super.getVendor() + '_' + super.getModel();
     }
 
     public String getVendor() {
-        return vendor;
+        return super.getVendor();
     }
 
     public String getModel() {
-        return model;
+        return super.getModel();
     }
 
     /**
