+++ specifications/active/phq8.yml
@@ -1,5 +1,8 @@
 #==================== Personal Health Questionnaire Depression Scale (PHQ-8) ======================#
 name: PHQ8
+vendor: PHQ
+model: 8
+version: v1
 assessment_type: QUESTIONNAIRE
 doc: Personal Health Questionnaire Depression Scale (PHQ-8) definition.
 data:
