+++ specifications/active/thincit.yml
@@ -1,4 +1,7 @@
 name: THINCIT
+vendor: LUNDBECK
+model: THINCIT
+version: v1
 assessment_type: APP
 doc: ThincIt Cognition in Depression app.
 data:
