+++ specifications/monitor/radar_prmt.yml
@@ -1,4 +1,7 @@
 name: ANDROID_PHONE
+vendor: ANDROID
+model: PHONE_MONITOR
+version: v1
 doc: Status updates from the RADAR-CNS pRMT app.
 app_provider: .application.ApplicationServiceProvider
 data:
