+++ specifications/passive/biovotion_vsm1.yml
@@ -1,6 +1,7 @@
 #===================================== Biovotion VSM1 Everion =====================================#
 vendor: BIOVOTION
 model: VSM1
+version: v1
 app_provider: .biovotion.BiovotionServiceProvider
 data:
   - type: ACCELEROMETER
