+++ specifications/passive/empatica_e4.yml
@@ -1,6 +1,7 @@
 #====================================== Empatica E4 Wristband =====================================#
 vendor: EMPATICA
 model: E4
+version: v1
 app_provider: .empatica.E4ServiceProvider
 data:
   - type: ACCELEROMETER
